+++ scripts/systems/SystemManager.mjs
@@ -11,6 +11,7 @@ import {SWADE} from "./handlers/SWADE.mjs";
 import {WFRP4e} from "./handlers/WFRP4e.mjs";
 import {Worldbuilding} from "./handlers/Worldbuilding.mjs";
 import {WorldsWithoutNumber} from "./handlers/WorldsWithoutNumber.mjs";
+import {DungeonCrawlClassics} from "./handlers/DungeonCrawlClassics.mjs";
 
 export class SystemManager {
   /**
@@ -25,6 +26,7 @@ export class SystemManager {
     CyberpunkRedCore,
     Demonlord,
     DND5e,
+    DungeonCrawlClassics,
     Dungeonworld,
     OldSchoolEssentials,
     Pathfinder2e,
@@ -67,6 +69,9 @@ export class SystemManager {
    */
   static #getSystemHandler() {
     const handler = this.#registeredHandlers[game.system.id];
+
+    if (!handler) return null;
+
     this.#instance = new handler();
 
     return this.#instance;